repeater: export IsMuted and add unit tests; keep ClearActive test helper
